+++ src/cell/tcell.rs
@@ -24,3 +24,11 @@ pub(crate) type ShareCellOwner = TCellOwner<ShareMarker>;
 pub(crate) fn new_share_cell_owner() -> ShareCellOwner {
     TCellOwner::new()
 }
+
+pub(crate) struct Share2Marker;
+pub(crate) type Share2Cell<T> = TCell<Share2Marker, T>;
+pub(crate) type Share2CellOwner = TCellOwner<Share2Marker>;
+
+pub(crate) fn new_share2_cell_owner() -> Share2CellOwner {
+    TCellOwner::new()
+}